move/rename the new classes
