+++ src/ca/cgjennings/apps/arkham/deck/Deck.java
@@ -3,7 +3,7 @@ package ca.cgjennings.apps.arkham.deck;
 import ca.cgjennings.apps.arkham.StrangeEons;
 import ca.cgjennings.apps.arkham.component.ComponentMetadata;
 import ca.cgjennings.apps.arkham.component.GameComponent;
-import ca.cgjennings.apps.arkham.component.conversion.ConversionContext;
+import ca.cgjennings.apps.arkham.component.ConversionContext;
 import ca.cgjennings.apps.arkham.deck.item.CardFace;
 import ca.cgjennings.apps.arkham.deck.item.Curve;
 import ca.cgjennings.apps.arkham.deck.item.DependentPageItem;
